requirements.txt
flask
python-dotenv
langchain
langchain_community
langchain_openai
faiss-cpu
pypdf
